More code updates
